+++ src/kernel_evm/CHANGES.md
@@ -34,7 +34,7 @@
 - Add support for `web3_sha3`. (!10112)
 - Add support for `eth_getUncleCountByBlockHash` and `eth_getUncleCountByBlockNumber`. (!10131)
 - Add support for `eth_getUncleByBlockHashAndIndex` and `eth_getUncleByBlockNumberAndIndex`. (!10134)
-
+- Add 2300 unit of gas for the RPC `eth_estimateGas` to account for the EIP2200. (!10176)
 
 ### Bug fixes